# api.yml
openapi: 3.0.3
info:
  title: API
  description: |
    API documentation including:
    - **Auth** – login, register, refresh (public); returns JWT for protected routes.
    - **User** – profile, badges, task history, streak, resume, profile pic (JWT required).
    - **Tasks** – get assigned tasks with **user_status** (completed, viewing, rejected, not_started); submit proof (JWT required).
    - **Leaderboard** – pan-india, state, college; each with optional period (all, weekly, monthly) or dedicated /weekly and /monthly routes; entry has name, rank, xp, profile_image, id, state, college.
    - **States** – list all states (public).
    - **Colleges** – list colleges by state (public).
  version: 1.0.0
servers:
  - url: /api
    description: API base path
security:
  - BearerAuth: []
paths:
...
  /tasks:
    get:
      summary: Get tasks (completed and ongoing)
      description: |
        Get all tasks assigned to the authenticated user. Each task includes **user_status** (completed, viewing, rejected, not_started) so you can show completed vs ongoing from one route.

        **Task submission flow:**
        - **not_started** → user has not submitted. Use POST /tasks/{id}/submit to submit proof.
        - **viewing** → user submitted; submission under review (DB pending). Cannot resubmit until approved or rejected.
        - **rejected** → submission rejected. User may resubmit via POST /tasks/{id}/submit if task deadline has not passed.
        - **completed** → submission approved. Task is done for the user; cannot resubmit.
      operationId: getTasks
      tags:
        - task
      responses:
        '200':
          description: List of tasks with user_status (completed | viewing | rejected | not_started)
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TaskWithUserStatus'
              example:
                - id: "880e8400-e29b-41d4-a716-446655440004"
                  title: "Complete your profile"
                  description: "Add your bio and profile picture"
                  xp: 25
                  type: "one_time"
                  proof_type: "image"
                  priority: "normal"
                  is_flash: false
                  is_weekly: false
                  created_by: "admin-uuid"
                  created_at: "2025-01-15T09:00:00Z"
                  status: "ongoing"
                  user_status: "completed"
                  submission_id: "770e8400-e29b-41d4-a716-446655440003"
                - id: "990e8400-e29b-41d4-a716-446655440005"
                  title: "Share post"
                  description: "Share our post on social media"
                  xp: 15
                  type: "one_time"
                  proof_type: "image"
                  priority: "normal"
                  is_flash: false
                  is_weekly: false
                  created_by: "admin-uuid"
                  created_at: "2025-01-16T09:00:00Z"
                  status: "ongoing"
                  user_status: "not_started"
                  submission_id: ""
                - id: "aa0e8400-e29b-41d4-a716-446655440006"
                  title: "Upload resume"
                  description: "Upload your resume"
                  xp: 50
                  type: "one_time"
                  proof_type: "image"
                  priority: "high"
                  is_flash: false
                  is_weekly: false
                  created_by: "admin-uuid"
                  created_at: "2025-01-17T09:00:00Z"
                  status: "ongoing"
                  user_status: "viewing"
                  submission_id: "bb0e8400-e29b-41d4-a716-446655440007"
        '401':
          description: Unauthorized
          content:
            text/plain:
              schema:
                type: string
              example: "Unauthorized"
        '500':
          description: Internal server error
          content:
            text/plain:
              schema:
                type: string
              example: "Failed to get tasks"
  /tasks/{id}/submit:
    post:
      summary: Submit task
      description: |
        Submit a task with proof file (image or video). Proof is uploaded to S3.
        Creates a submission with **status** `pending`. The task then appears with **user_status** `viewing` in GET /tasks (under review).

        **Resubmission:** Allowed only if the previous submission was **rejected** and the task deadline has not passed. On resubmit, the same submission row is updated (new proof, status back to `pending`).

        **Not allowed:** Already approved; or pending (under review); or rejected but task expired.
      operationId: submitTask
      tags:
        - task
      parameters:
        - name: id
          in: path
          required: true
          description: Task ID (UUID)
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - proof
              properties:
                proof:
                  type: string
                  format: binary
                  description: Proof file – image (JPG, PNG, GIF, WEBP) or video (MP4, MOV, AVI, MKV, WEBM), max 50MB
      responses:
        '201':
          description: Submission created or updated (resubmission). Status `pending`; task shows as `viewing` in GET /tasks.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Submission'
              example:
                id: "770e8400-e29b-41d4-a716-446655440003"
                task_id: "880e8400-e29b-41d4-a716-446655440004"
                user_id: "550e8400-e29b-41d4-a716-446655440000"
                proof_url: "https://bucket.s3.amazonaws.com/task-proofs/task-id/user-id_proof.jpg"
                status: "pending"
                created_at: "2025-01-20T14:00:00Z"
        '400':
          description: Bad request – task expired, not started, already submitted, or invalid file
          content:
            text/plain:
              schema:
                type: string
              examples:
                expired:
                  value: "Task has expired"
                not_started:
                  value: "Task has not started yet"
                already_approved:
                  value: "Task already approved. Cannot resubmit."
                pending:
                  value: "Task submission is pending review. Cannot resubmit."
                rejected_expired:
                  value: "Task has expired. Cannot resubmit rejected submission."
        '401':
          description: Unauthorized
          content:
            text/plain:
              schema:
                type: string
              example: "Unauthorized"
        '404':
          description: Task not found
          content:
            text/plain:
              schema:
                type: string
              example: "Task not found"
        '500':
          description: Internal server error
          content:
            text/plain:
              schema:
                type: string
              example: "Failed to upload proof file"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /auth/login
  schemas:
...
    Task:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        description:
          type: string
        xp:
          type: integer
        type:
          type: string
          description: Task type (e.g. one_time, recurring)
        proof_type:
          type: string
          description: Required proof type (e.g. image, video)
        priority:
          type: string
          description: Task priority (e.g. normal, high)
        start_at:
          type: string
          format: date-time
          nullable: true
        end_at:
          type: string
          format: date-time
          nullable: true
        is_flash:
          type: boolean
        is_weekly:
          type: boolean
        created_by:
          type: string
          format: uuid
        created_at:
          type: string
          format: date-time
        status:
          type: string
          enum: [ongoing, ended, completed]
          description: "ongoing = submission open (includes when user's submission was rejected, so they can resubmit); ended = time passed for submission (past end_at); completed = e.g. admin closed"
    TaskWithUserStatus:
      type: object
      description: Task plus user-specific completion status. Use user_status to show completed vs ongoing from GET /tasks.
      allOf:
        - $ref: '#/components/schemas/Task'
        - type: object
          properties:
            user_status:
              type: string
              enum: [completed, viewing, rejected, not_started]
              description: |
                completed = user has approved submission;
                viewing = submitted, under review (DB pending);
                rejected = submission rejected, may resubmit if task not ended;
                not_started = user has not submitted.
            submission_id:
              type: string
              description: Set when user has a submission (approved, viewing, or rejected); empty when not_started.
    Submission:
      type: object
      description: Task submission. When status is `pending`, the task shows as **user_status** `viewing` in GET /tasks.
      properties:
        id:
          type: string
          format: uuid
        task_id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        proof_url:
          type: string
          format: uri
        status:
          type: string
          enum:
            - pending
            - approved
            - rejected
          description: "pending = under review (viewing); approved = done; rejected = may resubmit if task not ended."
        admin_comment:
          type: string
          description: Admin comment (e.g. rejection reason); set on approve/reject.
        reviewed_by:
          type: string
          format: uuid
          description: Admin user ID who reviewed the submission.
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...